+++ components/src/main/java/org/patternfly/component/accordion/Accordion.java
@@ -16,6 +16,7 @@
 package org.patternfly.component.accordion;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -28,6 +29,8 @@ import org.patternfly.component.BaseComponent;
 import org.patternfly.component.ComponentType;
 import org.patternfly.component.HasItems;
 import org.patternfly.component.IconPosition;
+import org.patternfly.component.Ordered;
+import org.patternfly.component.list.ListItem;
 import org.patternfly.handler.ToggleHandler;
 import org.patternfly.style.Classes;
 import org.patternfly.style.Modifiers.Bordered;